refactor(airconv): split signature handling code
Also move the real instruction translation code into a dedicated translation unit.
