+++ src/airconv/meson.build
@@ -4,7 +4,9 @@ airconv_src = files([
  'air_signature.cpp',
  'air_operations.cpp',
  'dxbc_converter.cpp',
+ 'dxbc_converter_basicblock.cpp',
  'dxbc_instructions.cpp',
+ 'dxbc_signature.cpp',
  'metallib_writer.cpp'
 ])
